Feat/tutorial 3 traveloke (#6)
* feat/tutorial-3 update HotelModel.java

* feat/tutorial-3 create KamarModel

* feat/tutorial-3 create HotelDb, KamarDb, hotelservices

* feat/tutorial-3 create KamarService

* feat/tutorial-3 create HotelServiceImpl

* feat/tutorial-3 update HotelinMemoryService

* feat/tutorial-3 update hotelController

* feat/tutorial-3 create KamarController

* feat/tutorial-3 create home and form add hotel html

* feat/tutorial-3 create form add kamar html

* feat/tutorial-3 create form update hotel html

* feat/tutorial-3 create view hotel html

* feat/tutorial-3 create add kamar html

* feat/tutorial-3 final html

* feat/tutorial-3 final latihan

* feat/tutorial-3 final readme

* feat/tutorial-3 final

* feat/tutorial-3 final html

+++ traveloke/src/main/java/apap/tutorial/traveloke/controller/HotelController.java
@@ -2,50 +2,95 @@ package apap.tutorial.traveloke.controller;
 
 import java.util.List;
 
+import apap.tutorial.traveloke.model.HotelModel;
+import apap.tutorial.traveloke.model.KamarModel;
+import apap.tutorial.traveloke.service.KamarService;
+import apap.tutorial.traveloke.service.HotelService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-
-import apap.tutorial.traveloke.model.HotelModel;
-import apap.tutorial.traveloke.service.HotelService;
+import org.springframework.web.bind.annotation.*;
 
 @Controller
 public class HotelController {
+    @Qualifier("hotelServiceImpl")
     @Autowired
     private HotelService hotelService;
 
+    @Autowired
+    private KamarService kamarService;
+
+    @GetMapping("/")
+    private String home(){
+        return "home";
+    }
+
     //Routing URL yang diinginkan
     @RequestMapping("/hotel/add")
-    public String addHotel(
-        //Request parameter yang ingin dibawa
-        @RequestParam(value = "idHotel", required = true) String idHotel,
-        @RequestParam(value = "namaHotel", required = true) String namaHotel,
-        @RequestParam(value = "alamat", required = true) String alamat, 
-        @RequestParam(value = "noTelepon", required = true) String noTelepon,
+    public String addHotelFormPage(Model model){
+        model.addAttribute("hotel", new HotelModel());
+        return "form-add-hotel";
+    }
+
+    @PostMapping("/hotel/add")
+    public String addHotelSubmit(
+        @ModelAttribute HotelModel hotel,
         Model model
     ){
-
-        //Membuat objek HotelModel
-        HotelModel hotel = new HotelModel(idHotel, namaHotel, alamat, noTelepon);
-
-        //Memanggil service addHotel
         hotelService.addHotel(hotel);
-
-        //Add variabel id hotel ke "idHotel" untuk dirender pada thymeleaf
-        model.addAttribute("idHotel", idHotel);
-
-        //Return view template yang digunakan
+        model.addAttribute("idHotel", hotel.getId());
         return "add-hotel";
+    }
+
+    @GetMapping("/hotel/change/{idHotel}")
+    public String changeHotelFormPage(
+        @PathVariable Long idHotel, 
+        Model model
+    ){
+        try {
+            HotelModel hotel = hotelService.getHotelByIdHotel(idHotel);
+            model.addAttribute("hotel", hotel);
+            model.addAttribute("idHotel", hotel.getId());
+            return "form-update-hotel";
+        } catch (Exception e){
+            model.addAttribute("idProcess", idHotel);
+            return "cant-process";
+        }
+        
+    }
+
+    @PostMapping("/hotel/change")
+    public String changeHotelFormSubmit(
+        @ModelAttribute HotelModel hotel,
+        Model model
+    ){
+        HotelModel hotelUpdated = hotelService.updateHotel(hotel);
+        model.addAttribute("hotel", hotel);
+        return "update-hotel";
+    }
+
+    @GetMapping("/hotel/view")
+    public String viewDetailHotel(
+        @RequestParam(value="idHotel") Long idHotel,
+        Model model
+    ){
+        try {
+            HotelModel hotel = hotelService.getHotelByIdHotel(idHotel);
+            List<KamarModel> listKamar = kamarService.findAllKamarByIdHotel(idHotel);
+            model.addAttribute("hotel", hotel);
+            model.addAttribute("listKamar", listKamar);
+            return "view-hotel";
+        } catch (Exception e) {
+            model.addAttribute("idProcess", idHotel);
+            return "cant-process";
+        }
     }
 
     @RequestMapping("/hotel/viewall")
     public String listHotel(Model model){
         //Mendapatkan semua HotelModel 
-        List<HotelModel> listHotel = hotelService.getHotelList();
+        List<HotelModel> listHotel = hotelService.getHotelListSorted();
 
         //Add variabel semua HotelModel ke "listHotel" untuk dirender pada thymeleaf.
         model.addAttribute("listHotel", listHotel);
@@ -54,76 +99,23 @@ public class HotelController {
         return "viewall-hotel";
     }
 
-    @RequestMapping("/hotel/view")
-    public String detailHotel(
-        @RequestParam(value = "idHotel") String idHotel,
-        Model model){
-
-            //Mendapatkan HotelModel sesuai dengan idHotel 
-            HotelModel hotel = hotelService.getHotelByIdHotel(idHotel);
-
-            if(hotel!=null){
-                //Add variabel HotelModel ke "hotel" untuk dirender pada thymeleaf
-                model.addAttribute("hotel", hotel);
-
-                return "view-hotel";
-            } else {
-                model.addAttribute("idHotel", idHotel);
-                return "id-tidak-tersedia";
-            }
-    }
-
-    @GetMapping(value = "/hotel/view/id-hotel/{idHotel}")
-    public String detailPathHotel(
-        @PathVariable(value = "idHotel", required = true) String idHotel,
-        Model model
-    ){
-            //Mendapatkan HotelModel sesuai dengan idHotel 
-            HotelModel hotel = hotelService.getHotelByIdHotel(idHotel);
-
-            if(hotel!=null){
-                //Add variabel HotelModel ke "hotel" untuk dirender pada thymeleaf
-                model.addAttribute("hotel", hotel);
-
-                return "view-hotel";
-            } else {
-                model.addAttribute("idHotel", idHotel);
-                return "id-tidak-tersedia";
-            }
-    }
-
-    @GetMapping(value = "/hotel/update/id-hotel/{idHotel}/no-telepon/{noTelepon}")
-    public String updateNoTelepon(
-        @PathVariable(value = "idHotel", required = true) String idHotel,
-        @PathVariable(value = "noTelepon", required = true) String noTelepon,
-        Model model
-    ){
-        //untuk update nomor Telepon
-        HotelModel hotel = hotelService.updateNoTelepon(idHotel, noTelepon);
-
-        //Add variabel ke HotelModel untuk dirender pada thymeleaf
-        model.addAttribute("idHotel", idHotel);
-
-        if(hotel!=null){
-            return "update-notelp";
-        } else {
-            return "id-tidak-tersedia";
-        }
-    }
-
-    @GetMapping(value = "/hotel/delete/id-hotel/{idHotel}")
+    @GetMapping("/hotel/delete/{idHotel}")
     public String deleteHotel(
-        @PathVariable(value = "idHotel", required = true) String idHotel,
+        @PathVariable Long idHotel, 
         Model model
     ){
-        HotelModel hotel = hotelService.deleteHotel(idHotel);
-
-        model.addAttribute("idHotel", idHotel);
-
-        if(hotel!=null){
-            return "delete-hotel";
-        } else {
-            return "id-tidak-tersedia";
+        try{
+            List<KamarModel> kamar = kamarService.findAllKamarByIdHotel(idHotel);
+            model.addAttribute("idHotel", idHotel);
+            if(kamar.size()==0){
+                hotelService.deleteHotel(idHotel);
+                return "delete-hotel";
+            } else {
+                return "cant-delete";
+        }
+        } catch(Exception e){
+            model.addAttribute("idProcess", idHotel);
+            return "cant-process";
         }
     }
 }
